Create an aptly::mirror defined type
Per the class docs:

Create a mirror using `aptly mirror create`. It will not update, snapshot,
or publish the mirror for you, because it will take a long time and it
doesn't make sense to schedule these actions frequenly in Puppet.

It will also import the GPG key into the keyring that aptly defaults to.