Add hiera support using create_resources
With this addition it will be possible to use hiera directly to create repositories or mirrors. Example:
```
aptly::aptly_repos:
  myrepo:
    component: universe
```
The `require => Class['aptly']` had to be removed because this creates a circular dependency.
We now require the package and the config file instead since you may end up with repos
not in your configured rootdir if they are created before the config.