Pin versions in Gemfile
In 04defc241177c0b0e00298316b5ccae3e055e347 and d2161fcc3e37810f47e64abd467710c048430387
I updated the tests to reflect the version of the test framework that Travis was
running against.

Travis removes the `Gemfile.lock` before every run, which means it installs the
latest versions of these gems.

The correct way to fix this problem is to pin our gems to versions that we
know work with the syntax of our tests. This commit pins gems to the versions
which Travis is using (and passing) with and verifies that the tests pass
locally.