Add config_file parameter to repo define
If you're managing multiple aptly repositories on a single node you need
to be able to specify the config on a per-repo basis. This commit adds
the ability to pass a custom config for the repo commands, keeping the
existing default behavior. This code assumes you will be managing the
custom config files externally.