Make apt::source usage work with latest version
Since version 2.0.0 of the puppetlabs/apt module, the "key_server"
argument has been deprecated in favour of specifying a "key" hash. The
key_server parameter was removed completely in version 4.0.0 of the
module.

This changeset updates the apt::source resource block to use the new
style and bumps the supported version of the apt module up to include
the 4.x branch.